# A custom no-allocating queue for MemoryPoolBlock in the MemoryPool