@Scooletz 2016-10-06
@davidfowl Is there any benchmarking test suite I could use to write a custom test for the `MemoryPool`? I'd like to be able to prove the better performance instead of talking it through ;-)