@Scooletz 2016-10-06
It looks like the build has hung. The reason might be the number of allocated blocks which is quite high: `4096`. I know it's a lot but it's the easiest way to test the concurrent structure.