@benaadams 2016-10-06
Sure its not suffering from the [ABA problem](https://en.wikipedia.org/wiki/ABA_problem) on the CAS testing? (Which is one of the reasons the ConcurrentQueue uses segments which it discards)