@Scooletz 2016-10-07
It suffers from this. Segments might be dropped when the head reappeared as the head with additional values. I'm not closing a PR though, let me try to polish it.