@Scooletz 2016-10-08
The current approach will not suffer from ABA. I've updated the description to match implementation @benaadams . Still, it requires performance tests, so any guidance (is there already a perf suite in channels) would be greatly appreciated.