@benaadams 2016-10-12
Looking a bit closer, rather than changing `MemoryPool` you might want to add a new type that either inherits from `MemoryPool` or implements `IBufferPool`

Then in the perf tests you can compare `MemoryPool` to this version; the channels themselves [take an `IBufferPool`](https://github.com/davidfowl/Channels/blob/master/src/Channels/ChannelFactory.cs#L16-L23) so side-by-side implementations of the pools can coexist.