@Scooletz 2016-10-12
@benaadams I run the tests with the previous version (with the added benchmark) and then on the updated version. I did 3 runs each, selecting the best result. Let me present the results first to bring some numbers to the table.

#### Previous implementation

 Method |     Median |    StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------- |----------- |---------- |------ |------ |------ |------------------- |
 Leases | 64.0265 ns | 1.0048 ns | 89.00 | 67.00 | 24.00 |               5.76 |

#### New no-alloc implementation

 Method |     Median |    StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------- |----------- |---------- |------ |------ |------ |------------------- |
 Leases | 58.3106 ns | 1.0316 ns |     - |     - |  2.00 |               0.11 |