@Scooletz 2016-10-17
@benaadams @davidfowl I separated the `Poolable` class and compared that with a concurrent queue. It looks like there's not that much improvement in it beside removing all the allocations.